+++ src/main/java/com/example/tripit/user/repository/UserRepository.java
@@ -1,7 +1,6 @@
 package com.example.tripit.user.repository;
 
 import com.example.tripit.user.entity.UserEntity;
-import org.springframework.context.annotation.Bean;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -17,7 +16,7 @@ public interface UserRepository extends JpaRepository<UserEntity, Long>{
     UserEntity findByEmail(String email);
 
     @Query("SELECT u.userId FROM user u WHERE u.email = :email")
-    Integer findUserIdByEmail(@Param("email") String email);
+    long findUserIdByEmail(@Param("email") String email);
 
     //UserEntity findByNickname(String nickname);
